pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skinspire_v2"
version = "0.1.0"
description = "Hospital Management System"
requires-python = ">=3.12"
dependencies = [
    "Flask==3.1.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-Migrate==4.0.5",
    "cryptography==42.0.2",
    "PyJWT==2.8.0",
    "python-dotenv==1.0.1",
    "Werkzeug==3.1.3",
    "polars==0.19.19",
    "reportlab==4.0.8",
    "SQLAlchemy==2.0.36",
    "pytest==8.3.4",
    "pytest-mock==3.14.0",
    "schedule==1.2.2",
    "tabulate==0.9.0",
]

[tool.setuptools]
packages = ["app"]